S3: introduce static WandaModelAccess boundary
Replace dynamic adapter loading with a static model-access seam, split wanda API responsibilities into focused modules, migrate execution and extraction call sites, remove legacy adapter module, and update tests for deterministic Windows cleanup and new interfaces.
