+++ src/pywandahydra/execution/runner.py
@@ -61,7 +61,6 @@ def run(
     verbose: bool = False,
     workflow_name: str = "default",
     workflow_params: dict[str, Any] | None = None,
-    adapter_class: str = "pywandahydra.wanda.pywanda_adapter:PywandaAdapter",
 ) -> RunResult:
     """Run scenarios with the specified model and context.
 
@@ -80,7 +79,6 @@ def run(
         verbose: Enable detailed logging during execution.
         workflow_name: Post-processing workflow name.
         workflow_params: Post-processing workflow parameters.
-        adapter_class: Import path for the WandaAdapter implementation.
 
     Returns:
         Aggregated RunResult.
@@ -111,7 +109,6 @@ def run(
         run_root,
         workflow_name=workflow.name,
         workflow_params=workflow_params,
-        adapter_class=adapter_class,
     )
     if not plans:
         return RunResult(
