+++ src/pywandahydra/wanda/__init__.py
@@ -1,4 +1,4 @@
-"""WANDA integration package — adapter, session, API utilities."""
+"""WANDA integration package interfaces."""
 
-from .adapter import WandaAdapter  # noqa: F401
+from .model_access import WandaModelAccess  # noqa: F401
 from .session import wanda_session  # noqa: F401
